+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Releases
 - **Breaking** Removed ``thriftrw.wire.TType`` in favor of the
   ``thriftrw.wire.ttype`` module.
 - Performance improvements to ``BinaryProtocol`` implementation.
+- TypeSpecs for function arguments and results now have a reference back to the
+  function spec itself.
 
 
 0.5.2 (2015-10-19)